requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.5.3
bitarray==3.0.0
bitstring==4.3.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.1
dotenv==0.9.9
easyocr==1.7.2
ecdsa==0.19.0
esptool==4.8.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.0
h11==0.14.0
idna==3.10
imageio==2.37.0
intelhex==2.3.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyserial==3.5
pytesseract==0.3.13
python-bidi==0.6.6
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
reedsolo==1.7.0
scikit-image==0.25.2
scipy==1.15.2
shapely==2.0.7
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tifffile==2025.3.13
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.12.2
uvicorn==0.34.0
